+++ Dev/API-Server/src/main/java/org/dongguk/vsa/modeul/dialogue/repository/DialogueRepository.java
@@ -1,7 +1,7 @@
 package org.dongguk.vsa.modeul.dialogue.repository;
 
 
-import org.dongguk.vsa.modeul.dialogue.domain.Dialogue;
+import org.dongguk.vsa.modeul.dialogue.domain.mysql.Dialogue;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
