+++ Dev/API-Server/src/main/java/org/dongguk/vsa/modeul/file/domain/mongo/Component.java
@@ -1,4 +1,4 @@
-package org.dongguk.vsa.modeul.file.domain;
+package org.dongguk.vsa.modeul.file.domain.mongo;
 
 public interface Component {
     // FileLeaf와 DirecotryComposite에서의 공통 메소드를 정의하면 됩니다.
