+++ Dev/API-Server/src/main/java/org/dongguk/vsa/modeul/file/domain/mongo/FileLeaf.java
@@ -1,4 +1,4 @@
-package org.dongguk.vsa.modeul.file.domain;
+package org.dongguk.vsa.modeul.file.domain.mongo;
 
 import jakarta.persistence.Id;
 import lombok.AccessLevel;
