+++ Dev/API-Server/src/main/java/org/dongguk/vsa/modeul/file/repository/DirectoryCompositeRepository.java
@@ -1,6 +1,6 @@
 package org.dongguk.vsa.modeul.file.repository;
 
-import org.dongguk.vsa.modeul.file.domain.DirectoryComposite;
+import org.dongguk.vsa.modeul.file.domain.mongo.DirectoryComposite;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface DirectoryCompositeRepository extends MongoRepository<DirectoryComposite, String> {
