+++ Dev/API-Server/src/main/java/org/dongguk/vsa/modeul/file/repository/FileLeafRepository.java
@@ -1,6 +1,6 @@
 package org.dongguk.vsa.modeul.file.repository;
 
-import org.dongguk.vsa.modeul.file.domain.FileLeaf;
+import org.dongguk.vsa.modeul.file.domain.mongo.FileLeaf;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface FileLeafRepository extends MongoRepository<FileLeaf, String> {
