+++ Dev/API-Server/src/main/java/org/dongguk/vsa/modeul/modeullak/domain/mysql/ModeullakKeyword.java
@@ -1,18 +1,17 @@
-package org.dongguk.vsa.modeul.modeullak.domain;
+package org.dongguk.vsa.modeul.modeullak.domain.mysql;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.dongguk.vsa.modeul.tag.domain.Tag;
+import org.dongguk.vsa.modeul.keyword.domain.mysql.Keyword;
 
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "modeullak_tags")
-public class ModeullakTag {
-
+@Table(name = "modeullak_keywords")
+public class ModeullakKeyword {
     /* -------------------------------------------- */
     /* Default Column ----------------------------- */
     /* -------------------------------------------- */
@@ -22,6 +21,13 @@ public class ModeullakTag {
     private Long id;
 
     /* -------------------------------------------- */
+    /* Information Attribute ---------------------- */
+    /* -------------------------------------------- */
+    @Column(name = "content", length = 20, nullable = false)
+    private String content;
+
+
+    /* -------------------------------------------- */
     /* Many To One Mapping ------------------------ */
     /* -------------------------------------------- */
     @ManyToOne(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
@@ -29,15 +35,18 @@ public class ModeullakTag {
     private Modeullak modeullak;
 
     @ManyToOne(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
-    @JoinColumn(name = "tag_id", nullable = false)
-    private Tag tag;
+    @JoinColumn(name = "keyword_id", nullable = false)
+    private Keyword keyword;
+
 
     @Builder
-    public ModeullakTag(Modeullak modeullak, Tag tag) {
+    public ModeullakKeyword(String content, Modeullak modeullak, Keyword keyword) {
+        this.content = content;
         this.modeullak = modeullak;
-        this.tag = tag;
+        this.keyword = keyword;
 
-        modeullak.addModeullakTags(this);
-        tag.addModuellakTags(this);
+        modeullak.addModeullakKeywords(this);
+        keyword.addModeullakKeywords(this);
     }
+
 }
