+++ Dev/API-Server/src/main/java/org/dongguk/vsa/modeul/modeullak/repository/ModeullakKeywordsRepository.java
@@ -1,6 +1,6 @@
 package org.dongguk.vsa.modeul.modeullak.repository;
 
-import org.dongguk.vsa.modeul.modeullak.domain.ModeullakKeyword;
+import org.dongguk.vsa.modeul.modeullak.domain.mysql.ModeullakKeyword;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
