+++ Dev/API-Server/src/main/java/org/dongguk/vsa/modeul/user/repository/UserModeullakRepository.java
@@ -1,6 +1,6 @@
 package org.dongguk.vsa.modeul.user.repository;
 
-import org.dongguk.vsa.modeul.user.domain.UserModeullak;
+import org.dongguk.vsa.modeul.user.domain.mysql.UserModeullak;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
